Handle returning null for non-strings
