+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 
 yarn-error.log
+
+tests/out/test.wasm
